feature #14 : transactional retryable 분리
