+++ src/main/java/com/irum/productservice/domain/product/Internal/service/ProductStockService.java
@@ -1,0 +1,102 @@
+package com.irum.productservice.domain.product.Internal.service;
+
+import com.irum.global.advice.exception.CommonException;
+import com.irum.openfeign.dto.request.UpdateStockRequest;
+import com.irum.openfeign.dto.response.UpdateStockDto;
+import com.irum.productservice.domain.discount.domain.entity.Discount;
+import com.irum.productservice.domain.discount.domain.repository.DiscountRepository;
+import com.irum.productservice.domain.product.domain.entity.ProductOptionValue;
+import com.irum.productservice.domain.product.domain.repository.ProductOptionValueRepository;
+import com.irum.productservice.domain.product.domain.repository.ProductRepository;
+import com.irum.productservice.domain.store.domain.entity.Store;
+import com.irum.productservice.domain.store.domain.repository.StoreRepository;
+import com.irum.productservice.global.exception.errorcode.ProductErrorCode;
+import com.irum.productservice.global.exception.errorcode.StoreErrorCode;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class ProductStockService {
+    private final ProductOptionValueRepository productOptionValueRepository;
+    private final DiscountRepository discountRepository;
+    private final StoreRepository storeRepository;
+
+
+    @Transactional
+    public UpdateStockDto updateStockInTransaction(UpdateStockRequest request) {
+        Store store =
+                storeRepository
+                        .findByIdWithDeliveryPolicy(request.storeId())
+                        .orElseThrow(() -> new CommonException(StoreErrorCode.STORE_NOT_FOUND));
+        // 옵션 id 모으기
+        List<UUID> productOptionValueIdList =
+                request.optionValueList().stream()
+                        .map(UpdateStockRequest.OptionValueRequest::optionValueId)
+                        .toList();
+
+        // 옵션 조회. product group, product, store도 fetch join
+        List<ProductOptionValue> productOptionValueList =
+                productOptionValueRepository.findAllByIdWithFetchJoin(productOptionValueIdList);
+        Map<UUID, ProductOptionValue> povMap =
+                productOptionValueList.stream()
+                        .collect(
+                                Collectors.toMap(
+                                        ProductOptionValue::getId,
+                                        productOptionValue -> productOptionValue));
+
+        // 정합 점검 : 존재하지 않는 상품을 주문하고 있는지
+        validateAllOptionValuesExist(request, productOptionValueList);
+
+        // 재고 감소
+        for (UpdateStockRequest.OptionValueRequest optionValueRequest : request.optionValueList()) {
+            ProductOptionValue pov = povMap.get(optionValueRequest.optionValueId());
+
+            // 주문 검증 : 상점의 상품인지, 재고 부족 체크
+            validateStoreAndStock(pov, store, optionValueRequest);
+
+            pov.decreaseStock(optionValueRequest.quantity());
+        }
+
+        // 할인 조회
+        List<UUID> productIdList =
+                productOptionValueList.stream()
+                        .map(pov -> pov.getOptionGroup().getProduct().getId())
+                        .toList();
+        List<Discount> discountList = discountRepository.findAllByProductIds(productIdList);
+        Map<UUID, Integer> discountMap =
+                discountList.stream()
+                        .collect(Collectors.toMap(Discount::getId, Discount::getAmount));
+
+        return UpdateStockDto.from(store, productOptionValueList, discountMap);
+    }
+
+    /** 모든 옵션이 존재하는지 확인 */
+    private void validateAllOptionValuesExist(UpdateStockRequest request,
+                                              List<ProductOptionValue> povList) {
+        if (povList.size() != request.optionValueList().size()) {
+            throw new CommonException(ProductErrorCode.PRODUCT_NOT_FOUND);
+        }
+    }
+
+    /** 주문 검증 : 상점의 상품인지, 재고 부족 체크 */
+    private void validateStoreAndStock(ProductOptionValue pov, Store store, UpdateStockRequest.OptionValueRequest optionValueRequest) {
+        // 해당 상점의 상품을 주문하고 있는지
+        if (!pov.getOptionGroup().getProduct().getStore().getId().equals(store.getId())) {
+            throw new CommonException(ProductErrorCode.PRODUCT_NOT_IN_STORE);
+        }
+
+        // 재고보다 요청 물품 개수가 많을때
+        if (pov.getStockQuantity() < optionValueRequest.quantity()) {
+            throw new CommonException(ProductErrorCode.PRODUCT_OUT_OF_STOCK);
+        }
+    }
+}
